doc/groff.texi: Fix style nit.
Don't refer to the characters comprising the parameters to the .tr
requet as "arguments"; throughout the rest of this manual, request
"arguments" are separated by space characters.

Also characterize the translation target of a final, odd-parity input
character as a "fixed-width" space instead of an "unstretchable" one.

Some day we may have shrinkable/squeezable spaces and the existing
coinage was not future-proof.