doc/groff.texi: Fix terminological nit.
Reduce the number of locutions we have for referring to
gathered-but-not-yet-output text down to two.
	pending output line
	partially collected line
Ensure both are clearly discussed (the latter needed to be called out
for synonmy with the former, and given a Texinfo @dfn command).

Discard the following alternative constructions.
	current partial output line
	current partial line
	partially filled line
	current partially filled line
	current partial input line (yup)

The ones including the word "filled" are especially misleading because
they suggest that line collection is more closely coupled to filling
enablement than it is.

	.nf
	foo bar\c
	.mymacro
	baz qux
	.fi

In the above, the macro "mymacro" can observe a partially collected line
in spite of filling being disabled.