[docs]: Further improve intro to font files.
Motivation:

When I was first exposed to *roff many years ago, I knew so little that
the practice of referring to the inscrutably named "font files" "TR",
"HI", and so forth led me to believe that they were, somehow,  fonts.
That is, I thought that *roff did its own font rendering, and that these
weird files somehow provided enough information to do this, which didn't
seem plausible, but according to Unix fanboy culture, Murray Hill had a
room full of unimpeachable geniuses so how could I be sure?

It's much kinder to the reader to spend a word and say "description",
reinforcing the notion (which we should do more work in the "Fonts and
Symbols" node of our Texinfo manual to implant) that glyph rendering
remains the responsibility of the device.

A groff "font file" "AR" is simply not remotely the same sort of thing
that URWGothic-Book.t1 is, and we do readers and learners a grave
disservice by encouraging them to think that they are.

So I expect to be treating the term "font file" in our documentation
with some savagery henceforth, injecting the word "description" anywhere
I think there might otherwise be ambiguity.  I acknowledge that doing so
will be met with allergic reactions from Thompson disciples who find a
10-letter word abhorrent to read and even more so to write.