doc/groff.texi: Clarify no-break control char use.
Having looked at contrib/mm among other specimens, I see that even
capable macro package authors can be fuzzy on when the no-break control
character is meaningful and what, exactly, it does.

Attempt to elucidate this issue.  For most requests, the no-break
control character is indistinguishable in effect from the normal control
character.  It turns the "br" request into a no-op.  For eleven
others, it delays but does not cancel the request's effect.