[docs]: Eschew term "no-fill mode".
It is awkward (even if impressively pedigreed--"nofill" without the
hyphen, explained as "non-fill", dates back at least to the 1976 edition
of CSTR #54).

Say instead that "filling" is "enabled" or "disabled" as appropriate.

* doc/groff.texi:
* doc/meintro.me:
* doc/meref.me:
* doc/pic.ms:
* man/groff.7.man:
* man/roff.7.man:
* src/roff/troff/troff.1.man:
* tmac/groff_man.7.man.in:
* tmac/groff_mdoc.7.man:
* tmac/groff_me.7.man:
Do it.

I did not update uses of "{no-?,}fill mode" in change logs, TODOs, or
code comments (except in groff_man*(7) which is already filthy with my
handprints anyway), or in mom, which has a carefully curated lexicon and
is well-maintained by Peter Schaffter.

Also...

* doc/groff.texi (Manipulating Filling and Adjustment): Add explicit
  notice of synonymy of "fill mode" with "filling of output lines" in
  description of `fi` request and concept index entries for those
  accustomed to the older terminology.

  (Line Control): Tighten discussion of `\c` escape by adding cross
  reference to the node above.

  (Warnings): Recast description of `break` warning category per the
  above and explicitly note the .l register.

* man/groff.7.man (Requests/Request short reference) <fi, nf>: Sync with
  node "Manipulating Filling and Adjustment" from our Texinfo, adding
  notice of effect on .u register.

* src/roff/troff/troff.1.man (Warnings) <break>: Sync with node
  "Warnings" from our Texinfo manual, noting relevance of .l register.