[grog]: Refactor even more aggressively.
Refactor grog even more aggressively.  Make the tedious search for
subs.pl unnecessary by inlining it into grog.pl directly.  This makes
the script stand alone and much more convenient to work with, and
doesn't contribute to bloat, in my opinion--subs.pl was 10 times the
size of grog.pl[1], and most of the code in the latter was dedicated to
trying to locate subs.pl.

* src/roff/grog/grog.am (GROG, dist_grog_DATA): Drop variables.

  (grog): Remove target's dependency on `$(GROG)`.

* src/roff/grog/grog.pl: Inline most of subs.pl (except for
  author/license comment banenr).  Delete sub-hunting logic.

* src/roff/grog/subs.pl: Delete.

[1] git says: "rename src/roff/grog/{subs.pl => grog.pl} (97%)"