tmac/e.tmac-u: Move comment.
This is to make it easier to merge the handful of later changes to the
original "me" package that occurred after groff forked its own version.

BSD maintenance of "me" slowed to a trickle by 1991 and, confusingly,
the package got a new version number and copyright date in 1993 (for
4.4BSD) with no other changes to the actual file content.  Software
copyright and version notices are as scrupulous as Hollywood accounting.