doc/groff.texi: Improve examples.
I've long been bothered by the TeX logo examples I added; to fit them
within the margins of the print version of out Texinfo manual I had to
reduce the precision of the movements by quite a bit.  The idea of
propagating such a crude mockery of Knuth's logo filled me with horror.

So, change it.

* doc/groff.texi (Strings): Improve examples.
  (Page Motions): Add trailing '\"' to TeX logo example that _does_ fit.
  Add footnote to explain its presence by cross reference.