[ms]: Document cautionary note about .in better.
An ms "reset" is only meaningful to macro package source divers and the
advice that was given implied that it's up to the user to figure out
which macros cause `par@reset` (not even identified by name) to be
called.  Try to be a little more generous to the general user.