[grog]: Drop filename extension-based inference.
* src/utils/grog/grog.pl (handle_file_ext): Delete.

  (top level): Drop call site.

I feel this subroutine was poorly judged.  Many of the filename
extensions it uses are seldom if ever seen outside of this file and the
groff_filenames(5) man page (also by Bernd).  I also regard the project
of the subroutine as exceeding grog's proper scope, which is to _make
heuristic inferences based on the contents on an input stream_.

Admittedly, this does regress one of the smoke tests; doc/webpage.ms is
no longer detected as an ms(7) document.  It's an interesting case,
loaded with groff_www(7) macros; all of the ms macros it calls happened
to have names that man(7) _also_ uses.  Perhaps this why Bernd resorted
to file name extensions, but I have another idea.