[docs]: Apply lexical discipline to "adjustment".
Update discussions and code comments to reserve the term "adjustment" to
the only situation where the formatter logic actually
applies--"full justification" or "normal"/"both" margins in *roff
parlance.  Use forms of the term "align" for contexts where on the left
or right margin is expected to be straight.  The overloading of the term
"adjustment" in roff systems is, I think, a historical artifact of there
once being only two alternatives: adjustment to both margins, and no
adjustment (with a presumption of left-to-right orthography).

We can't rename the .ad request, but we can stop abusing terminology
when we discuss it or macros that interact with it.