groff_mm(7): Generalize claims about mm.
In a couple of cases the compatibility wrapper prefix was used in
self-reference to mm, but regarding claims that are true about mm
implementations generally.  It's not accurate to make such claims
particular to groff ms.

This also lowers the impact of the compatibility wrapper feature, the
removal of which has been discussed in
<https://savannah.gnu.org/bugs/?59425>.