roff(7): Clarify a minor historical point.
One can argue about the semantics of "fonts" versus "styles" on
teletypewriters, but one thing is clear: the C/A/T had a symbol font
whereas the Teletype Model 37 did not (if you wanted the special
non-ASCII character set, you had to embed SI and SO ASCII escapes in the
input stream yourself[1]).  So recast a statement to be less arguable.

[1] See the kbd(7) page from the First Edition Unix manual on page 13 of
    <https://www.bell-labs.com/usr/dmr/www/pdfs/man71.pdf>.