[libgroff, troff]: Add debug diagnostic level.
* src/include/error.h: Declare functions `debug_with_file_and_line` and
  `debug`.
* src/libs/libgroff/error.cpp: Add `DEBUG` to enum `error_type`.

  (do_error_with_file_and_line): Add case for `DEBUG` in switch.

  (debug, debug_with_file_and_line): Add new functions.

Do the same for troff since it has a private implementation of the
diagnostic functions (thanks to `output_warning()`).

* src/roff/troff/input.cpp: Add `DEBUG` to enum `error_type`.

  (do_error_with_file_and_line): Add case for `DEBUG` in switch.

  (debug, debug_with_file_and_line): Add new functions.

There are no call sites for debug() and I don't plan for there to be; I
am adding these for developer convenience because it's the second time
I've had to do so and I told myself after the first (ripping them back
out after they'd helped me solve a problem) that if I ever needed them
again I'd add them permanently.