[ms]: Encourage use of more intuitive feature.
Even reading the original thread that led to the fractional point size
rescaling feature[1], I find it hard to motivate usage of it when the
"p" scale unit is available and other implementations (like Heirloom)
produce no diagnostics when given a register setting or point size they
can't interpret (they just ignore the request).

So why do this:
  .\" I want 10.5 point type if I can get it.
  .ie \n(GS .nr PS 10500
  .el       .nr PS 10
when you can do this?
  .\" I want 10.5 point type if I can get it.
  .ie \n(GS .nr PS 10.5p
  .el       .nr PS 10

[1] https://lists.gnu.org/archive/html/groff/2004-01/msg00014.html