glilypond(1): Fix synopsis solecism.
"-" is (unless preceded by a "--" option) a special value of a "file"
operand, not really something with its own distinct semantic value
relative to other command parameters.  While I'm sure the point could be
argued, none of groff's other man pages for commands that support this
ancient notation for a standard file stream give it annotation in the
synopsis in this way.  So get rid of it.

Also relocate argument-free short option (-k) to the front of the
synopsis, as is done in our other pages.