groff_me(7): Sync copyright, license with 4.4BSD.
...and the William Hoskins advertising clause rescission.

See <https://svnweb.freebsd.org/csrg/share/me/me.7?view=log>.

Not much changed in this man page after groff forked it circa 1990.

The SCCS identifier claimed version 6.4 but the typo fix in that
revision documented at the above URL was not present.  It looks like
what we really had was revision 6.3.  Here's the subsequent history of
the file.

6.4	Grammar fix.  I independently made the same fix (among many
	others) in commit 23b606c7, 3 Nov 2017.
6.5	Copyright/license update.
6.6	"[M]ake groff happy".  This looks like "sed 's/^\.ti0/.ti 0/'".
	It is inapplicable and irrelevant to our version of the macro
	reference, which uses a tbl(1) table.
8.1	Copyright notice update, adding year 1993.