groff(7): Fix content nits.
Improve explanation of units.  Drop "actual" from description of basic
unit for output device (there's neither such a thing as a "virtual"
output device, and the Denglish use of "actual" for "current" is also
inapplicable because the output device cannot change during the lifetime
of the formatter).  Identify "v" unit as "vee" (like "em" and "en").