[man]: Tidy up interface between an-footer and BT.
[man]: Clean up the interface between `an-footer` and `BT`, and HTML
header and footer suppression generally.  Move responsibility for
checking the `ps4html` register from the latter to the former.  This is
not something a `BT` redefiner {see groff_man(7)} should have to worry
about.

* tmac/an.tmac (BT): Drop test of `ps4html` register and early return.

  (an-header, an-footer): Return immediately if
  `an-suppress-header-and-footer` is true.  Remove conditional on
  `an-is-output-html`.

  (initialization): Define `an-suppress-header-and-footer` register;
  true if `an-is-output-html` is true, or if `ps4html` is defined.

I observe that (AIUI) were it not for the page-number dependence of the
ps4html method of getting page content into HTML output, we could use
continuous rendering mode for HTML.