[troff]: Enable ASCII in device control escapes.
[troff]: Convert special character glyphs corresponding to Unicode Basic
Latin ("ASCII") code points to those code points when they occur in
device escapes.  (They should be correct for IBM code page 1047 as well,
but this is untested.)  This is necessary for encoding URLs in device
control commands.  Special character identifiers are presumed to be the
defaults documented in groff_char(7); this is a design gap that we
should consider addressing.  (We don't have a way to ask "is this the
special character corresponding to Unicode basic Latin code point X?")

* src/roff/troff/input.cpp (encode_char): Do it.

I'm not documenting this in NEWS as it feels like a pretty dusty corner
even though I'm about to leverage it for something of much higher
visibility.