doc/meref.me: Fix content nit.
The -me Reference Manual is fairly scrupulous about documenting the
default values of its registers, but $v and $V are two exceptions.
[This commit] documents the default values for these (both 120).

    $ groff -me <<<'.tm \n($v \n($V'
    120 120

There is also what I feel is an oddity in the nearby wording: "The line
spacing as a percentage of the point size expressed in units is stored
in \n($v."  I'm not sure what the phrase "expressed in units" is
supposed to communicate; it's not clear whether it modifies "line
spacing" or "point size," and neither really makes sense, even mentally
adding the presumably intended "basic" before "units."

    $v is a percentage; if it meaningfully has a unit at all, that unit
    is %.

    In what units the point size is expressed or stored is irrelevant to
    a percentage: 120% of the point size is not dependent on the units
    expressing that point size.

Thus this patch removes the phrase entirely.  If someone has a better
idea what that phrase is trying to say, it could instead be reworded to
say it more clearly.

Fixes <https://savannah.gnu.org/bugs/index.php?61314>.

[It would be nice to say "type size" instead of "point size", but
consistent usage would demand a much more intrusive change.  --GBR]