[sboxes]: Restructure.
Drop article's quotation of its own source.  Maintenance of the document
in this form will be much too tedious; instead we can generate it.

We mark two locations in the document to bring this off: first, where a
footnote is used within the document but remains "live" in the
quoted-source form (and produces different footnote text within and
outside the quotation), and second, the place where the quoted source
itself will be interpolated.