[docs]: Fix "number" style nit.
Our documentation is pretty loose with the term "number", using it even
when a more specific term that implies cardinality or ordering would
help the reader to reason about the concept being discussed.  Try to
improve matters.

* The registers `.R`, `.ce`, `.rj` store, and the request `hlm` takes,
  a count of objects.
* Say "quantity" or "index" instead of "number" in some contexts as
  well.

(The distinction between measurement and counting could probably be
better recognized in some places, too.)