doc/groff.texi: Recast `ig` request description.
Shift the caveat about interpolation of auto-incrementing registers to a
footnote.  Stop suggesting that there is anything special about GNU
troff's handling of it.  Inform reader of optional whitespace in control
line ending region.  Update example to use an end macro and thereby
illustrate that the end macro call is interpreted normally, not ignored.