[grohtml]: Add assertion.
* src/devices/grohtml/post-html.cpp (html_printer::do_file_components):
  Add assertion.

The command "echo | ./build/test-groff -Thtml" was core dumping on me.
`current_paragraph` was a null pointer.  groff 1.22.4 was fine.  A
bisect did not track down the problem, and it no longer manifested after
return to Git HEAD and rebuilding.  I presume there is a missing
dependency somewhere that skipped necessary rebuilding/relinking.  Until
that can be tracked down, explicitly slam on the brakes if the situation
arises again.