[tbl]: Accept `\&` as an empty table entry.
* src/preproc/tbl/table.cpp (table::add_entry): Suppress diagnostic
  about non-empty table entries classified as `_` or `=` if the entry
  consists of exactly `\&`.

* src/preproc/tbl/tbl.1.man (Table data): Document this idiom.

This behavior change is intended as an ergonomic aid for those who use
the default tab character and whose text editors don't make tabs visible
(as can be done with Vim's 'list' option).  It is consistent with other
tbl and *roff usage of this token, which does not represent anything
visible or printable.