[me]: Fix ugly, incorrect use of "\(aa" in docs.
An acute accent is not an apostrophe.  Someone should travel back in
time and strike Berkeley with acute bolts of lightning repeatedly to
emphasize the grave nature of their error.

I'm leaving the occurrence in meref.me's "Special Characters and Marks"
section uncorrected because that whole section needs a rewrite
(pending).