doc/meref.me: Undocument `_M` register.
This reverts part of d9ba1bd9, 5 August 2004.

This is the only me(7) register we document that is named with a leading
underscore, and thus lies outside the naming convention described in the
introduction to this document.  If authors of me(7) documents need to
test this register, they can still do so, and if a more "official"
interface is required, then my opinion is that we should expose a string
cacheing the first argument to `++` and name it `$T` or `$t` (both are
available).

Also update documentary comment in tmac/e.tmac to align terminology with
forthcoming commit.