[docs]: Update cross references to 'me' ref man.
Alter the typography to quote this article-length document instead of
underlining or italicizing it.  It has also lost its leading
hyphen/dash.

Thanks to Dave Kemper for the reminder to do this.

Also change the formatter reference to groff's Texinfo manual instead of
CSTR #54, which was further referred to under its 1976 (not 1991) title.