doc/meref.me.in: Relocate material.
* Move `m[1234]` macros from "Headers and Footers" to "roff Support".
* Move `ep` macro from "Headers and Footers" to "Miscellaneous".

These seem like more appropriate locations, and further I think it is
helpful for newcomers to have to absorb less material to learn about
header and footer manipulation features.

`ep` is explicitly described as being applied to pending footnote
flushes, and not to do with footers (it in fact does _not_ suppress the
footer), so again a relocation seems appropriate.