groff_mmse(7): Stop sourcing localization package.
Man pages should not do this.  It will mess up any subsequent man pages
rendered in the same formatter run that use a different language.  This
has always been discouraged (except by this dubious example), but was
only noticeable if you paid attention to the hyphenation breaks being
wrong in subsequent pages.  Now that we have localizations for the
manual section titles, it's more obvious.

There isn't a way to batch-render man pages in multiple languages while
being agnostic about this fact on the command line[1] unless we were to
prescribe adding an ".mso" request for the man page's intended
localization file to _every_ page, including English ones, and that
seems like a non-starter due to the breadth of the impact, the status of
the `mso` request as a groff extension, and the groff-specific names for
the localization packages.

Possibly, few people apart from me care about reliable batch-rendering.

[1] It can be done, with gnosis, by interleaving the localization macro
    files among a set of man page operands.  Assuming a default language
    of English, consider the following hypothetical arrangement.
      $ groff -man od.1 tmac/fr.tmac fr/ls.1 fr/cat.1 tmac/en.tmac tar.1