doc/groff.texi: Fix notation error.
It's pretty misleading to leave the closing delimiter off of a \X device
control escape sequence.  Supply it, and add a roman ellipsis to
indicate the presence of elided content.

Also, mark additional nodes as reviewed for correct grave accent and
apostrophe usage in examples, including the foregoing, which now renders
with the correct glyphs.