[docs]: Be more careful with control characters.
...when they are word-initial in filled text lines.

The idea is to prevent baffling surprises if a sentence is recast in a
text editor, and a word beginning with a dot or neutral apostrophe is
reflowed onto the beginning of the next input line--suddenly it is
parsed as a control character.

Fixes <https://savannah.gnu.org/bugs/?51312>.

Some of these edits permit other solecisms to persist.  I'll fix them
separately.