[docs]: Improve terminological discipline.
Migrate from "base line" to "baseline".  Also drop "baseline" from
injection in the midst of the "vertical spacing" term (which is defined
in terms of the baseline).

Add "baseline" and "vertical spacing" as terms in the at present
notional "Concepts" section of roff(7), which the page has long promised
but not yet delivered.