[ms]: Recast overview/document structure stuff.
Make the discussion of tables of contents and indices a bit less feeble.
Much of it may not remain long where it is; we could use a real
discussion of the `IX` macro and a bit of user tutorial about handling
the tasks of generating these.

Also bring our 3(!) copies of this material into sync.