[tbl]: Add unit test for format-time diagnostics.
Ensure that the various format-time diagnostics that a tbl-preprocessed
document can generate (1) occur when they are supposed to and (2) are
appropriately suppressed by the "nowarn" and/or "nokeep" region options.
One of them is an error, not a warning, and cannot be suppressed the
former, but _can_ be by "nokeep", which is arguably a bug
(Savannah #61878).

* src/preproc/tbl/tests/format-time-diagnostics-work.sh: Do it.
* src/preproc/tbl/tbl.am (tbl_TEST): Run test.

Test fails at this commit because one scenario[1] results in two
diagnostic messages for a single problem; fix pending.

[1] This one:

.pl 3v
.ll 10n
.TS
tab(;) expand;
L L.
abcde;fghij
.TE