Fix clarity problems and doubled words.
...and some recent typos.

* ChangeLog: Recast my own entry regarding a post-1.22.4 refactor of the
  man(7) font style alternation macros, for greater attempted clarity
  and more normative use of terminology.  Also be less vague about what
  'tty.tmac' does with the 'latin1' and 'cp1047' macro files.
* ChangeLog.119, NEWS (1.06[!]): Fix doubled word typos.
* ChangeLog: Fix recent typo "easile", extraneous phrase "not a
  warning", and inappropriate plural.

I recognize that some people feel it is inappropriate to modify older
content of ChangeLog and NEWS files in any way.  Permit me to paraphrase
myself in response[1] to Ingo Schwarze.

---
> changing historical information in the ChangeLog and the NEWS file
> would be wrong.

I agree that misrepresenting history would be wrong, but correcting
solecisms like doubled words is not a misrepresentation of history--it's
a _more accurate_ representation of history.

Because my own brain hates me, it is only after committing and pushing
that I am likely to notice egregiously bad typos or missing words in my
ChangeLog entries.

Material misrepresentations of fact in such files should either be
corrected in-place or supplemented with "[recte: ]" or similar
annotation.

There is the question of how much time has to pass before we consider
that such documents (or what portions of them, in the case of "NEWS")
have stabilized such that the bracketed-annotation approach is more
appropriate than the former.  The question does not arise in this case.

Anyway, like you, I don't exactly welcome further use of the Savannah
system for this sort of thing.  It would be better for such "linty"
patches to go to the groff@ mailing list, and better still for someone
to contribute a style checker that we can run automatically as part of
builds to catch these things automatically so that a developer can fix
them before they're even introduced into a pushed commit.

What ever became of the Writer's Workbench, eh?

As I recall, a doubled-word finder for plain text that works across line
boundaries is an exercise in the O'Reilly _sed & awk_ book.  It would be
a nice thing for a new developer to contribute!  :)
---

[1] https://savannah.gnu.org/bugs/?51062