[man]: Fix manual section title localization.
Commit c64fd60dff was not quite cooked.  The localized strings (in
English) for the manual section titles need to be set at every document
load (i.e., when `TH` is called), but only if the groff locale is
"english".  When switching locales, these strings were getting "stuck"
in their non-English translations even when returning to an English
page.

* tmac/an.tmac (an*localize): New macro initializes strings.

  (TH): Call `an*localize` if the locale is "english".

  (initialization): Drop initialization of section title strings.

Continues fixing <https://savannah.gnu.org/bugs/?61734>.

Tested with several scenarios, including this one that starts out with a
non-English page and then switches back and forth between languages.
The package is not tripped up and the center header remains appropriate.

$ ./build/test-groff -man -msv EXPERIMENTS/swedish.1 \
  tmac/en.tmac EXPERIMENTS/english.1 \
  tmac/sv.tmac EXPERIMENTS/swedish.1 \
  tmac/en.tmac EXPERIMENTS/english.1