[man pages]: Drop unnecessary use of \& escapes.
Use of the non-printing input break \& _before_ a dot is a recommended
practice to prevent the dot from becoming misinterprted as a control
character when a text line is reflowed/refilled in a text editor.

However, this reasoning does not apply to control lines, which generally
cannot be reflowed without damaging the document.

So get rid of 'em.