[man pages]: Fix style nits.
Permit trailing non-quotation punctuation after metasyntactic variable
names to be in italics.  These are less likely to acquire semantic tags
in the future than file names or command names are, because they are far
less likely to be copied and pasted, and they typeset better this way
(though they look slightly weirder on underlining terminals).