doc/groff.texi: Relocate material.
...to discuss colors right after "Sizes" and before "Strings".

* Update menu to reflect new ordering.
* Update GNU Info node navigation links.

As I noted to Dave in Savannah #51072, 'the "Colors" section/node seems
to come much later than it could.  It should be relocated earlier, no
later than "Sizes".  (In fact, nearly everything that an environment
saves/preserves should be discussed by the point that environments are
introduced, with exceptions only for arcane necessities like the
enablement status of compatibility mode or similar.)  The color-related
escapes sequences and requests do obvious things; apart from the mild
complication of color channels and spaces, they are straightforward to
present and use.'