doc/groff.texi: Rename an index.
...from "Font File Keyword Index", to "File Keyword Index".

We've been at pains lately to say "font _description_ file", and
furthermore this index applies to 'DESC' file keywords as well.
However, "Device and Font Description File Keyword Index" is much too
long a chapter title for the typeset version of this manual.  Shorten it
far enough to avoid an ugly break.