[sboxes]: Affix package name to early diagnostics.
* sboxes.tmac: Prefix early diagnostics with package file name
  literally.  (We don't yet know if we can use \*[] string
  interpolation.)  It is better to violate DRY a little here than to
  obscure from the user what is emitting the diagnostic.

We arguably shouldn't do this for the main macro packages because we
don't know what file names they will be installed under, due to the
compatibility wrappers feature (see the NEWS entry for groff 1.22.4).
In those cases it is better to omit the suggestion of a file name
producing the diagnostic than to risk uttering an incorrect one, which
can madden a person attempting to troubleshoot problems.

But we can do it for other packages unique to groff...