doc/groff.texi: Revise discussion of `\B' escape.
Clarify that it interpolates a value.  Add example of usage.  Help
reader avoid having to leap forward to a discussion of register formats.

Clarify: it is not the "entire expression" that needs to be surrounded
by parentheses to permit use of spaces, just the (sub)expression that
contains them.

Add example of (formally) complex numeric expressions.