doc/groff.texi (Identifiers): Revise.
Note colors as a category of identifier.

Visually tighten presentation of invalid input characters by moving it
from an itemized list to ordinary paragraphs.

Move discussion of what is done with invalid input characters in
identifiers to precede the laundry list, uninteresting to most readers,
of invalid code points in supported ISO and EBCDIC encodings.

Add giant footnote confessing the validity of some control characters in
identifiers, explain why it was done in the past, and discourage it.

Convert examples of valid identifiers from a spacey list to a regular
paragraph, and update them to be more illustrative.

Add examples of how to cope with identifiers beginning with '(' or '['.

Use @samp instead of @code for sample input.