[sboxes] fix non-portable syntax in the msboxes.pdf target
The DOC_GROFF facility uses the $< variable which POSIX only defines
for use in suffix rules, but the msboxes.pdf rule is not a suffix rule
and uses it anyway.  Besides, the whole point of DOC_GROFF is to
manipulate standard input, but then providing a file name argument
afterwards causes groff to discard the modified input.

Branden pointed out that we do not want a .ms.pdf suffix rule here
because the particular command is very specialized and other future
conversions from .ms to .pdf are unlikely to need -msboxes but
likely to need other macros or preprocessors.  So this commit uses
my alternative solution of simply using DOC_GROFF_ONLY in place of
DOC_GROFF.

Feedback and OK from gbranden@.