doc/groff.texi: Note significant historical point.
"One of the most amusing and unexpected consequences of phototypesetting
was the Unix standard error file (!).   After phototypesetting, you had
to take a long wide strip of paper and feed it carefully into a smelly,
icky machine which eventually (several minutes later) spat out the paper
with the printing visible.

One afternoon several of us had the same experience -- typesetting
something, feeding the paper through the developer, only to find a
single, beautifully typeset line: "cannot open file foobar"   The
grumbles were loud enough and in the presence of the right people, and a
couple of days later the standard error file was born..." -- Steven C.
Johnson