ChangeLog: Remove misleading statement.
The deleted text was a relic of an early, never-pushed, and ultimately
insufficient approach to solving the problem.

Since we now bail out of abbreviating strings that contain potentially
recursive escape sequences (more importantly for our purposes,
"non-atomic" ones as handled by the `substring` request), we _could_
continue to trim them in the middle as done before, but I developed
second thoughts about the desirability of doing so.